SysInfo.Drasil

System Information

data System where

Data structure for holding all of the requisite information about a system to be used in artifact generation.

Constructors

SI

Fields

:: (CommonIdea a, Idea a, Idea b, Quantity e, Eq e, MayHaveUnit e, Quantity h, MayHaveUnit h, Quantity i, MayHaveUnit i, HasUID j, Constrained j)

=> { _sys :: a

, _kind :: b

, _authors :: People

, _purpose :: Purpose

, _background :: Background

, _scope :: Scope

, _motivation :: Motivation

, _quants :: [e]

, _instModels :: [InstanceModel]

, _datadefs :: [DataDefinition]

, _configFiles :: [String]

, _inputs :: [h]

, _outputs :: [i]

, _constraints :: [j]

, _constants :: [ConstQDef]

, _sysinfodb :: ChunkDB

, _usedinfodb :: ChunkDB

} -> System

Lenses

class HasSystem c where

Minimal complete definition

system

Methods

system :: Lens' c System

authors :: Lens' c People

background :: Lens' c Background

configFiles :: Lens' c [String]

constants :: Lens' c [ConstQDef]

datadefs :: Lens' c [DataDefinition]

instModels :: Lens' c [InstanceModel]

motivation :: Lens' c Motivation

purpose :: Lens' c Purpose

scope :: Lens' c Scope

sysinfodb :: Lens' c ChunkDB

usedinfodb :: Lens' c ChunkDB

Reference Database

type Purpose = [Sentence]

Project Example purpose.

type Background = [Sentence]

Project Example background information, used in the What section of README.

type Scope = [Sentence]

Project Example scope.

type Motivation = [Sentence]

Project Example motivation.

citeDB :: System -> BibRef

Helper for extracting a bibliography from the system information.

Utility Helper Functions

ccss :: [Sentence] -> [ModelExpr] -> ChunkDB -> [DefinedQuantityDict]

Gets a list of defined quantities (DefinedQuantityDicts) from Sentences and expressions that are contained in the database (ChunkDB).

ccss' :: [Sentence] -> [ModelExpr] -> ChunkDB -> [QuantityDict]

Gets a list of quantities (QuantityDicts) from Sentences and expressions that are contained in the database (ChunkDB).

vars :: ModelExpr -> ChunkDB -> [QuantityDict]

Gets a list of quantities (QuantityDict) from an equation in order to print.