SysInfo.Drasil.GetChunk

Description

Utilities to get grab certain chunks (from Expr, Sentence, etc) by UID and dereference the chunk it refers to.

Documentation

ccss :: [Sentence] -> [ModelExpr] -> ChunkDB -> [DefinedQuantityDict]

Gets a list of defined quantities (DefinedQuantityDicts) from Sentences and expressions that are contained in the database (ChunkDB).

ccss' :: [Sentence] -> [ModelExpr] -> ChunkDB -> [QuantityDict]

Gets a list of quantities (QuantityDicts) from Sentences and expressions that are contained in the database (ChunkDB).

combine :: Sentence -> ChunkDB -> [DefinedQuantityDict]

Combines the functions of vars and concpt to create a list of DefinedQuantityDicts from a Sentence.

getIdeaDict :: Sentence -> ChunkDB -> [IdeaDict]

Gets a list of ideas (IdeaDict) from a Sentence in order to print.

vars :: ModelExpr -> ChunkDB -> [QuantityDict]

Gets a list of quantities (QuantityDict) from an equation in order to print.