SysInfo.Drasil

System Information

data SystemInformation where

Data structure for holding all of the requisite information about a system to be used in artifact generation.

Constructors

SI

Fields

:: (CommonIdea a, Idea a, Idea b, HasName c, Quantity e, Eq e, MayHaveUnit e, Quantity f, MayHaveUnit f, Concept f, Eq f, Quantity h, MayHaveUnit h, Quantity i, MayHaveUnit i, HasUID j, Constrained j)

=> { _sys :: a

, _kind :: b

, _authors :: [c]

, _purpose :: d

, _quants :: [e]

, _concepts :: [f]

, _instModels :: [InstanceModel]

, _datadefs :: [DataDefinition]

, _configFiles :: [String]

, _inputs :: [h]

, _outputs :: [i]

, _defSequence :: [Block SimpleQDef]

, _constraints :: [j]

, _constants :: [ConstQDef]

, _sysinfodb :: ChunkDB

, _usedinfodb :: ChunkDB

, refdb :: ReferenceDB

} -> SystemInformation

data Block a

for listing QDefinitions in SystemInformation.

Constructors

Parallel a [a]

sysinfodb :: Lens' SystemInformation ChunkDB

Reference Database

data ReferenceDB

Database for internal references. Contains citations and referrable concepts.

type RefMap a = Map UID (a, Int)

Database for maintaining references. The Int is that reference's number. Maintains access to both num and chunk for easy reference swapping between number and shortname/refname when necessary (or use of number if no shortname exists).

citeDB :: SystemInformation -> BibRef

Helper for extracting a bibliography from the system information.

rdb :: BibRef -> [ConceptInstance] -> ReferenceDB

Smart constructor for creating a reference database from a bibliography and concept instances.

simpleMap :: HasUID a => [a] -> RefMap a

Constructor that makes a RefMap from things that have a UID.

citationDB :: Lens' ReferenceDB BibMap

conceptDB :: Lens' ReferenceDB ConceptMap

Utility Helper Functions

ccss :: [Sentence] -> [ModelExpr] -> ChunkDB -> [DefinedQuantityDict]

Gets a list of defined quantities (DefinedQuantityDicts) from Sentences and expressions that are contained in the database (ChunkDB).

ccss' :: [Sentence] -> [ModelExpr] -> ChunkDB -> [QuantityDict]

Gets a list of quantities (QuantityDicts) from Sentences and expressions that are contained in the database (ChunkDB).

combine :: Sentence -> ChunkDB -> [DefinedQuantityDict]

Combines the functions of vars and concpt to create a list of DefinedQuantityDicts from a Sentence.

getIdeaDict :: Sentence -> ChunkDB -> [IdeaDict]

Gets a list of ideas (IdeaDict) from a Sentence in order to print.

vars :: ModelExpr -> ChunkDB -> [QuantityDict]

Gets a list of quantities (QuantityDict) from an equation in order to print.